Entity permissions backend

+++ elements/entity/entity.router.ts
@@ -4,7 +4,6 @@ import mongoose from "mongoose";
 import ConnectedRequest from "../../globalTypes/ConnectedRequest";
 import ResponseDto from "../../globalTypes/ResponseDto";
 import PaginationResponse from "../../globalTypes/PaginationResponse";
-import superAdminProtectMiddleware from "../../middleware/superAdminProtectMiddleware";
 import protectMiddleware from "../../middleware/protectMiddleware";
 import EntityCreateCommand from "./dto/EntityCreateCommand";
 import EntityReadDto, { toReadDto } from "./dto/EntityReadDto";
@@ -14,19 +13,27 @@ import EntityUpdateCommand from "./dto/EntityUpdateCommand";
 import EntitiesGetCommand from "./dto/EntitiesGetCommand";
 import { IUser } from "../user/user.model";
 import EntitiesSearchCommand from "./dto/EntitiesSearchCommand";
+import roleService from "../role/role.service";
+import { StaticPermission } from "../entityPermission/entityPermission.model";
+import entityRepository from "./entity.repository";
 
 const router = Router();
 
 router.post(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, EntityCreateCommand, any>,
     res: Response<ResponseDto<EntityReadDto>>
   ) => {
     const command: EntityCreateCommand = req.body;
     const currentUser: IUser = req.user;
+
+    roleService.checkEntityPermission({
+      user: currentUser,
+      staticPermission: StaticPermission.Create,
+      modelId: req.body.modelId.toString(),
+    });
 
     const entity: IEntity = await entityService.createEntity(
       command,
@@ -43,13 +50,19 @@ router.post(
 router.put(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, EntityUpdateCommand, any>,
     res: Response<ResponseDto<EntityReadDto>>
   ) => {
     const command: EntityUpdateCommand = req.body;
     const currentUser: IUser = req.user;
+
+    roleService.checkEntityPermission({
+      user: currentUser,
+      staticPermission: StaticPermission.Update,
+      modelId: req.body.modelId.toString(),
+    });
+
     const entity: IEntity = await entityService.updateEntity(
       command,
       currentUser
@@ -64,12 +77,19 @@ router.put(
 
 router.post(
   "/getEntitiesByModel",
+  protectMiddleware,
   async (
     req: ConnectedRequest<any, any, EntitiesGetCommand, any>,
     res: Response<ResponseDto<PaginationResponse<EntityReadDto>>>
   ) => {
     const command: EntitiesGetCommand = req.body;
     const { entities, total } = await entityService.getEntitiesByModel(command);
+
+    roleService.checkEntityPermission({
+      user: req.user,
+      staticPermission: StaticPermission.Read,
+      modelId: command.modelId.toString(),
+    });
 
     return res.status(200).send({
       success: true,
@@ -84,13 +104,25 @@ router.post(
 router.delete(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, mongoose.ObjectId[], any>,
     res: Response<ResponseDto<void>>
   ) => {
     const entitiesIds: mongoose.ObjectId[] = req.body;
-    await entityService.deleteEntities(entitiesIds);
+
+    if (entitiesIds.length > 0) {
+      const entity: IEntity = await entityRepository.getById(
+        entitiesIds[0].toString()
+      );
+
+      roleService.checkEntityPermission({
+        user: req.user,
+        staticPermission: StaticPermission.Read,
+        modelId: entity.model._id.toString(),
+      });
+
+      await entityService.deleteEntities(entitiesIds);
+    }
 
     return res.status(200).send({
       success: true,
@@ -106,6 +138,12 @@ router.post(
     res: Response<ResponseDto<PaginationResponse<EntityReadDto>>>
   ) => {
     const command: EntitiesSearchCommand = req.body;
+
+    roleService.checkEntityPermission({
+      user: req.user,
+      staticPermission: StaticPermission.Read,
+      modelId: command.modelId.toString(),
+    });
 
     const { entities, total } = await entityService.search(command);
